pyproject.toml
[project]
name = "crypto-dashboard"
version = "0.1.0"
description = ""
authors = [
    {name = "ling",email = "200826704@qq.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "eventlet (>=0.40.3,<0.41.0)",
    "django-redis (>=6.0.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]

[tool.black]
line-length = 100  # 设置每行最大字符数
include = '\.pyi?$'  # 指定要格式化的文件类型
exclude = '''
/(
  | .git
  | .venv
  | build
  | dist
)/
'''  # 排除不需要格式化的目录

[tool.isort]
profile = "black"  # 确保isort与black的格式兼容
line_length = 100  # 建议与black的line-length保持一致
multi_line_output = 3  # 设置多行导入时的输出风格
include_trailing_comma = true  # 在多行导入的最后一个条目后加逗号